+++ documentation-website/src/client/constants/versions.js
@@ -3,13 +3,13 @@ export default {
   "addon-active": "1.0.0-beta.5",
   "addon-ancestors": "1.0.0-beta.5",
   "addon-prefetch": "1.0.0-beta.6",
-  core: "1.0.0-beta.23",
-  mobx: "1.0.0-alpha.0",
-  react: "1.0.0-beta.17",
-  "react-native": "1.0.0-alpha.2",
-  redux: "1.0.0-beta.2",
-  "side-effect-scroll": "1.0.0-beta.6",
+  core: "1.0.0-beta.24",
+  mobx: "1.0.0-alpha.1",
+  react: "1.0.0-beta.18",
+  "react-native": "1.0.0-alpha.3",
+  redux: "1.0.0-beta.3",
+  "side-effect-scroll": "1.0.0-beta.7",
   "side-effect-title": "1.0.0-beta.8",
   svelte: "1.0.0-beta.4",
-  vue: "1.0.0-beta.14"
+  vue: "1.0.0-beta.15"
 };