+++ packages/react/types/CuriBase.d.ts
@@ -1,8 +1,8 @@
 /// <reference types="react" />
-import React from 'react';
-import PropTypes from 'prop-types';
-import { CuriRouter, Response, Navigation } from '@curi/core';
-import { CuriContext } from './interface';
+import React from "react";
+import PropTypes from "prop-types";
+import { CuriRouter, Response, Navigation } from "@curi/core";
+import { CuriContext } from "./interface";
 export interface CuriBaseProps {
     router: CuriRouter;
     render: (r: Response, navigation: Navigation, c?: CuriRouter) => React.ReactElement<any>;