+++ packages/react/types/Curious.d.ts
@@ -1,8 +1,8 @@
 /// <reference types="react" />
-import React from 'react';
-import PropTypes from 'prop-types';
-import { CuriContext } from './interface';
-import { CuriRouter, Response, Navigation } from '@curi/core';
+import React from "react";
+import PropTypes from "prop-types";
+import { CuriContext } from "./interface";
+import { CuriRouter, Response, Navigation } from "@curi/core";
 export interface CuriousProps {
     render(p: CuriousRenderProps): React.ReactElement<any>;
     router?: CuriRouter;