+++ packages/vue/types/interface.d.ts
@@ -1,9 +1,9 @@
-import { CuriRouter, Response, Navigation } from '@curi/core';
+import { CuriRouter, Response, Navigation } from "@curi/core";
 export interface ReactiveResponse {
     response: Response;
     navigation: Navigation;
 }
-declare module 'vue/types/vue' {
+declare module "vue/types/vue" {
     interface Vue {
         $curi: ReactiveResponse;
         $router: CuriRouter;