+++ packages/vue/types/plugin.d.ts
@@ -1,5 +1,5 @@
-import { PluginObject } from 'vue';
-import { CuriRouter } from '@curi/core';
+import { PluginObject } from "vue";
+import { CuriRouter } from "@curi/core";
 export interface CuriPluginOptions {
     router: CuriRouter;
 }